refactor: Move <SelectInput/> to @chainlit/components
