+++ frontend/src/components/organisms/playground/editor/MessageWrapper.tsx
@@ -9,7 +9,7 @@ import {
   Typography
 } from '@mui/material';
 
-import SelectInput from 'components/organisms/inputs/selectInput';
+import { SelectInput } from '@chainlit/components';
 
 import { playgroundState } from 'state/playground';
