+++ libs/components/src/inputs/index.tsx
@@ -1,4 +1,7 @@
 export { InputLabel } from './InputLabel';
 export { InputStateHandler } from './InputStateHandler';
 export { TextInput } from './TextInput';
+export { SelectInput } from './selects/SelectInput';
+
+export type { SelectInputProps, SelectItem } from './selects/SelectInput';
 export type { TextInputProps } from './TextInput';
